fix(api, robot-server, shared-data): inject run metadata into `FileProvider` (#19851)
Works towards EXEC-1978

When FileProvider is instantiated, it captures a bound method reference to FileProviderExecutor's callback. Any subsequent state changes to the FileProviderExecutor instance won't be reflected when the callback executes, since the method binding happened at instantiation time. To fix, let's just keep the run-related state on the FileProvider, since the updated instance of FileProvider is injected into PE on each run.