feat(shared-data): add wasteChuteLocation
This is a new kind of legacy location added in
https://github.com/Opentrons/opentrons/pull/19684 (
b801f7f5fda47a36a7b9fbc869fa002a9701a8bc ) to present better information
when a protocol tries to use a labware that has been thrown away. Add a
type binding for it in the LabwareLocation type that PE load commands
can take and emit, and some ts type predicates for figuring out when
it's present.