feat(robot-server): Add HTTP API support for retrieving image metadata for a run (#19853)
Works towards EXEC-1978

To support the camera image gallery in the app, we will need two endpoints: one that supplies camera image file metadata for a particular run, and one that allows the client to get the actual image blob.

This commit addresses the former. We add a /dataFiles/{runId}/images resource which provides a list of ImageFileMetadata, containing all the data the app cares about for the image gallery.