feat(robot-server): support auto deleting generated data files (#19858)
Closes EXEC-1986

This commit adds support for automatically deleting output data files when a run record "ages out" of the settings-defined max run record limit.

The RunAutoDeleter is injected with the DataFileAutoDeleter. The DataFileAutoDeleter deletes the data files associated with a run that is marked for deletion by the RunAutoDeleter.

The DataFileAutoDeleter uses the following strategy when cleaning up data files for a particular run:

- Always delete all entries from output_data_files table for the run.
- Always delete all physical files in the images directory for this run, since they are currently never used as input for a run (we can always revisit this behavior if that changes).
- For each output file associated with a run: if it has no entries in input_data_files table, fully remove from data_files_table and delete the physical file in the data file directory.
- For each output file associated with a run: if a file has entries in input_data_files table, preserve the physical file (so the ones located in dat_dir/file_id/file) and keep entries in the input_data_files table and the data_files table.