refactor(app): rename getAllDefinitions.ts `getAllDefinitions()` -> `getAllLatestDefs()` (#19869)
# Overview

This is part 1 in a long chain of PRs to fix tiprack lookups in PD.

Yes, at first glance, the code in this PR has nothing to do with PD. But
`app/src/local-resources/labware/utils/getAllDefinitions.ts` defines
this function called `getAllDefinitions()` that contrary to its name,
does **not** return all labware definitions. Instead, it filters the
list to only return the latest definitions.

This has nothing to do with PD. But it's getting in my way because there
is **another** function in `shared-data/js/labware.ts` called
`getAllDefinitions()` that **does** return all labware definitions. I'm
about to do some refactoring that will look very confusing if we have 2
functions with the same name.

So this first PR renames `getAllDefinitions()` to `getAllLatestDefs()`.

This will probably all make sense in about 10 more PRs.

## Test Plan and Hands on Testing

Global search for all the references to the renamed function.
Run the unit tests in CI.

## Risk assessment

Low, so far.