fix: handle wasteChuteLocation in frontend (#19860)
This is a new kind of legacy location added in
https://github.com/Opentrons/opentrons/pull/19684 (
b801f7f5fda47a36a7b9fbc869fa002a9701a8bc ) to present better information
when a protocol tries to use a labware that has been thrown away. Add a
type binding for it in the LabwareLocation type that PE load commands
can take and emit, and some ts type predicates for figuring out when
it's present.

This is a whole lot of code across the entire FE. it's validated by ts
but I need to add tests throughout and do some manual tests.

The PR is organized so each commit is a different subsystem, which might
be a nicer way to read it.

## Todo
- [x] update unit tests
- [x] test manually with associated protocol (anything that throws a
labware away)

Closes RQA-4772