fix: misspellings of "configuration" (#19875)
# Overview

We misspelled the word "configuration" a lot. Fortunately that's what
find and replace is for.

## Test Plan and Hands on Testing

Automated tests only.

## Changelog

Find and replace across full monorepo. This appears to affect strings
only, not functional code.

## Review requests

Is it a no-no to go back and fix this in old schema versions? It doesn't
affect functionality as far as I know. But we can omit those if it would
even possibly cause problems.

## Risk assessment

low