feat(protocol-designer): create and wire up tip selection for OT-2 (#19885)
This PR adds PipetteShadow components and correct deck svg for OT-2
tiprack and tip selection. It also fixes a step navigation issue for
OT-2 moveLiquid and mix tools

Note that there is a bug in the OT-2 single-channel shared pipe type
definitions such that their SVGs have 0 height and 0 width. This will be
addressed separately.