refactor(shared-data): simplify and fix caching in getAllDefinitions() (#19898)
# Overview

Part 3 of a chain of PRs to eventually fix tiprack lookups in PD.

We attempted to implement memoization in `getAllDefinitions(blockList:
string[])` by caching the result to the global variable `_definitions`,
but the implementation is incorrect and dangerous because there's only
one `_definitions`, and once it's set, we'll keep returning it for all
calls to `getAllDefinitions()` even if you specify a different
`blockList` later (or no `blockList` at all).

This PR removes the caching entirely.

And I changed the blocklists `LABWAREV2_DO_NOT_LIST` and
`PD_DO_NOT_LIST` into `Set`s. This means that we can check if any
particular labware is in the blocklist in roughly O(1) time, so we don't
have to feel bad about removing the caching. (I.e., building the return
value for `getAllDefinitions()` would take about as long as calling
`Object.values(...)` or `Object.entries(...)` on the definitions list,
and we do those things *all the time*, so if we're OK with that, we
should be OK with an uncached `getAllDefinitions()`.)

Furthermore, if you look at the files that use `getAllDefinitions()`,
such as `labware-library/src/definitions.tsx`, they often implement
**their own** layer of caching, so we really shouldn't feel bad about
removing the caching from `getAllDefinitions()` to make it more correct.

## Test Plan and Hands on Testing

Ran `yarn vitest protocol-designer/ step-generation/ app/`.
Will run CI tests.

## Risk assessment

Low.