refactor(shared-data,app-protocol-designer): getAllLabwareDefs() -> getAllDefinitions() (#19899)
# Overview

Make all of Max's dreams come true ...

Part 4 in a chain of PRs for tiprack definition lookups in PD.

After PR #17631, we ended up with 2 functions for getting the labware
definitions:
- `getAllLabwareDefs()` returned all the standard labware definitions.
- `getAllDefinitions(blockList)` returned all the standard labware
definitions except the ones in `blockList`, and also implemented
caching, but not quite correctly.

That PR left us with a TODO to combine the two functions.

We've fixed the caching issue now, so we can now go ahead and combine
the two functions. This PR eliminates `getAllLabwareDefs()` and changes
all the callers to call `getAllDefinitions()` instead.

## Test Plan and Hands on Testing

Ran `vitest` locally.
Ran CI tests.

## Risk assessment

Low-ish.