feat(api, robot-server): Error on disk space violations (#19909)
Closes EXEC-1963

Whenever a protocol command attempts to generate a data file, we should check if:

The overall disk space is sufficient.
If it's an image generating command, that the images directory disk space is sufficient.
If either of these conditions are violated, we should throw a StorageLimitReachedError.

Note that we deprecate the hard file limit error we were using before. This is true for the absorbance reader file output as well - now we exclusively check disk space.