refactor(shared-data): getAllLabwareDefs() -> getAllDefinitions() for new code in `edge` (#19911)
# Overview

In the PD 8.6.1 branch, we replaced `getAllLabwareDefs()` with
`getAllDefinitions()`, satisfying this old TODO:
https://github.com/Opentrons/opentrons/blob/238586e8376170e27b751198be78c749f75498ba/shared-data/js/labware.ts#L71-L75

However, @rclarke0's PR #19088 introduced some new calls to
`getAllLabwareDefs()` in `edge` that are not in the PD 8.6.1 branch. So
this PR migrates those calls too to unbreak `edge`.

## Test Plan and Hands on Testing

Hopefully the CI tests will pass now.

## Risk assessment

Low-ish.