refactor(app): Deprecate `outputFileIds` usage (#19917)
Closes EXEC-2004

The app currently assumes the outputFileIds from the run record comprises entirely of data file ids that are CSV files, but that is no longer the case. These assumptions lead to unintended behaviors like rendering download links for each individual image file, and trying to parse jpegs as CSVs. We have a hook for providing descriptive metadata about each data file id, so let's create a utility hook that uses that metadata to filter for data file ids that are exclusively csv file ids and use it.