chore(labware-designer): Fix reference to undefined `_NODE_ENV_` constant (#19928)
## Overview

This fixes a broken `labware-designer` build. It was whitescreening with
an error relating to `_NODE_ENV_` not being defined.

## Details

`labware-designer` does not actually use `_NODE_ENV_` anywhere. However,
some of its dependencies do. Those dependencies' configs try to define
`_NODE_ENV_` for themselves, so things *ought* to work. But
`labware-designer` includes those dependencies with a simple Vite import
alias—naively including their source files as if they were part of
`labware-designer` itself—so the dependencies' project configs are
bypassed completely, and `_NODE_ENV_` is never defined.

The quick fix is for `labware-designer`'s Vite config to define
`_NODE_ENV_` even though it shouldn't need to.

We've fixed other occurrences of this problem twice before: once in
#19586, and once in a PR of @mjhuff's that I'm having trouble digging
up. So this PR adds loud `# todo` comments to all of the config files to
try to stop this from getting any worse.