feat(app, components): add low storage warnings (#19926)
Closes EXEC-2008

Adds an "image storage is low" warning on both the desktop and ODD with slightly different behavior depending on the app.

The /health endpoint contains relevant information about the current disk space of the images directory and general system disk space. When either the images directory is considered too big or the system general disk space is considered too small, a protocol command that generates an output file will fail. At the time of this commit, those limits are 2GB and 250MB respectively.

The app effectively takes these robot-server default limits and adds a buffer to them, if that buffer limit is violated for either the images directory or system disk space, we render the low space banner.