fix(protocol-designer): Reordering steps should count as modifying the protocol (#19942)
In Protocol Designer, when you click the "Create new" button to create a
new protocol, it warns you if that would take you away from a protocol
that has unsaved changes.

Its mechanism for tracking that involves a hand-coded list of all the
things that count as modifying a protocol. That list was missing
`REORDER_STEPS` and `REORDER_SELECTED_STEP`. So if you changed the order
of steps in the timeline and did nothing else, it wouldn't warn you when
you clicked "Create new."

Closes AUTH-2438.