fix(api): dynamic pipetting bug fixes (#19958)
# Overview
Found a few slight issues when testing.

Couldn't set a flow rate below max because we didn't update the
flow_rate vs rate change for dynamic pipetting and it was trying to go
at flow_rate^2

Mix was coming out of the liquid between aspirate and dispense because
the normal waypoint motion planner for aspirate was doing a square move.
Now it does a force direct move between points.

Slight problem with setting the last location in pipette core.