fix(app): dont use live commands for module setup
Live commands don't really work that well in general and really really
don't work that well for modules, which have their state synced to the
engine from hardware once when they're loaded (since the engine
otherwise is the thing that's in control of them). Since we're creating
a maintenance run for module setup anyway, let's use that maintenance
run to run the module prep commands.

How to do this is a little trickier; the prep commands want to run
before we display the UI to put the calibration adapter in place (since
most of the time the thing you're prepping is the module's ability to
take the adapter, such as opening the thermocycler lid). We can do this
with a gross run-once hook in the appropriate component, but that
run-once machinery goes away when the component is unmounted... which it
is when you go backwards in the flow... and then we try and load
everything again and fail. Luckily, with a little more interaction with
the robot and knowledge of some peculiarities of the engine, we can get
that squared away by using a known labware ID, and bob's your uncle.

Closes RQA-2020