docs: OT2 backmatter (#19681)
# Overview

This PR updates and revises the backmatter sections for the OT-2 manual.
It should clearly cover topics similar to the Flex manual.

Sandbox:
https://sandbox.docs.opentrons.com/docs-ot2-revisions-backmatter/

RTC-864

## Changelog

3 new files replace OT-2 ch. 11, Appendix. You only have to review
these:

- `support.md`
- `open-source.md`
- `additional-docs.md`

This also significantly changes the current OT-T end chapter. That
version is an assorted grab-bag of info about Github, our warranty,
unrelated stuff about liquid handling, and just alot of "no, I can't
even..." sort of stuff.

While not a verbatim duplicate of the Flex manual, it is very similar.
Removed some Flex-specific things that aren't relevant to the OT-2.
Still, very boilerplate-ish.

## Review requests

Anything from the current OT-2 manual that's missing and should be in
here? We're going in with a heavy hand.

## Risk assessment

Low, docs; moderate because of the large, installed OT-2 base.

---------

Co-authored-by: Ed Cormany <edward.cormany@opentrons.com>