add mac to the connectivity csv section of robo-diag (#19971)
# Overview

Adds an output of the robots WiFi and Ethernet MAC addresses to the
Robot diagnostics QC script.

Customers have requested the ability to know the MAC address of their
robot before they receive it or set it up. By logging the MAC address in
the output of the diagnostic scripts we can now look up the MAC address
of a robot based on it's serial number in our production database.

## Test Plan and Hands on Testing

Ran the test script on a Flex in the office. Checked that the MAC
addresses were contained in the output CSV.

## Changelog

Modifies the connectivity section of the flex diagnostics QC script in
hardware testing to add wifi and ethernet mac address outputs.

## Review requests



## Risk assessment

Minimal only impacts hardware testing code

Co-authored-by: Ryan Howard <ryan.howard@opentrons.com>