refactor(protocol-designer): call getMatchingTipLiquidSpecsFromSpec() with tiprackDef (#19901)
# Overview

Part 5 for tiprack definition lookups in PD.

Back in PR #19041, @ncdiehl11 introduced the function
`getMatchingTipLiquidSpecsFromSpec()` that takes a tiprack URI and tries
to look up the tiprack definition. It sometimes fails, producing Sentry
errors like:
- https://opentrons-76.sentry.io/issues/6912737526/?project=4509363266387968
- https://opentrons-76.sentry.io/issues/6912737527/?project=4509363266387968

However, there is no need for `getMatchingTipLiquidSpecsFromSpec()`
itself to look up the tiprack definition, because the caller already has
the tiprack definition.

So this PR changes `getMatchingTipLiquidSpecsFromSpec()` to take a
tiprack labware definition instead of a tiprack URI.

Note that the existing implementation in `load-file/migration/8_5_0.ts`
to find the tiprack definition is not quite correct, and would fail for
custom tipracks. But we'll fix that in a future PR.

## Test Plan and Hands on Testing

CI tests.

## Risk assessment

Low.

---------

Co-authored-by: Nick Diehl <47604184+ncdiehl11@users.noreply.github.com>