docs(github): simplify pull request template, stop checking in HTML comments (#20015)
# Overview

Despite my admonishments, we continue to check in commits that begin
with:
```
<!--
Thanks for taking the time to open a Pull Request (PR)! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:
...
```
This is noisy and distracting, especially when I examine the git log
from the command line. (Github hides HTML comments when you preview the
text via the web UI, but the HTML comments are included verbatim in the
git commit.)

This change deletes the introductory text, and changes all the other
instructions in the template from comments into body text.

I'm hoping that this will make authors delete the instructions as they
fill in each section, because if they don't, the instructions will show
up very obviously in Github and in the email messages.

## Risk assessment

Low.