refactor(shared-data,app,pd): unify LabwareDefinitionsByURI definitions (#20020)
# Overview

We've ended up with many definitions for the type "labware definitions
by URI", including:
- `LabwareDefinitionsByURI` in `getLabwareDefinitionsByURIForProtocol.ts`
- `LabwareDefinitionsByUri` in `getLoadedLabwareDefinitionsByUri.ts`
- `LabwareDefByDefURI` in `protocol-designer`

Due to re-exports, the first 2 are exported at the top level of
`@opentrons/shared-data`.

I suspect that when writing code, people just click on whichever pops up
first in their IDE autocomplete. They would probably be surprised at
which file their import is coming from.

This PR merges the first 2 definitions and moves them to
`shared-data/js/types.ts`. The 3rd definition is a PD-specific alias for
`LabwareDef2ByDefURI`, and this PR changes it to a proper alias.

(Note that there's also `RunLoadedLabwareDefinitionsByUri`, which I'm
not going to touch for now because its name is a less tempting target
for autocomplete.)

## Test Plan and Hands on Testing

Run CI.

## Risk assessment

Low.