fix(robot-server, api-client, app): Populate filename with server-generated value (#20021)
Closes RQA-4813

In the desktop app's image gallery, there is an overflow menu that enables users to download a single image. Clicking that option renders an OS download dialogue and populates the filename with an app-generated filename, which follows a Design-specified naming convention.

However, generating the filename on the app fails to capture the PAPI image_capture command's ability to specify a prefix to the naming convention via the filename param. The filename exists in its entirety on the server in the data files table, so let's just bubble that filename up through a few data file metadata endpoints instead of requiring the app to attempt to reflect the behavior of the server.