refactor(js): Prefer `value!` syntax over `value as T`, when possible (#20032)
## Overview

This implements a discussion we had in-person a few weeks ago. It makes
a small improvement to type safety and it resolves two lint rules that
were fighting each other.

## Changelog

If the type of `value` is `T | null` or `T | undefined`, then TypeScript
gives us two ways to cast it to `T`:

1. `value as T`
2. `value!`

We had one lint rule that was warning if you wrote `value as T` when you
could have just written `value!`. That was in conflict with an unrelated
lint rule that was erroring if you wrote `value!` at all.

If we have to choose between the two styles, `value!` is better because
it makes it impossible to accidentally cast to a completely unrelated
type, like going from `number` to `string`.

So:

* It's now an *error,* not a warning, to write `value as T` when you
could have just written `value!`.
* It's now a *warning,* not an error, to write `value!` at all.

Also:

* It's now a *warning,* not an error, to write a chain of `?.` followed
by `!`, like `a?.b?.c!`.

I think this was a good error ([see its
docs](https://typescript-eslint.io/rules/no-non-null-asserted-optional-chain/))
and this is an unfortunate downgrade. The change from `value as T` to
`value!` incidentally got eslint to notice a lot of these errors for the
first time, and we can't fix them without risking a a change to runtime
behavior. (Removing `?.` is a change to runtime behavior.) So I left
them alone for now.

## Risk assessment

Low. There shouldn't be anything in here that affects runtime behavior.