fix(protocol-designer): hydrate Form.tipRack to support custom tipracks (#20034)
# Overview

(This is Part 6 of tiprack definition fixes for PD.)

The problem we're trying to solve:
`Hydrated{MoveLiquid/Mix}FormData.tipRack` contains the URI of a
tiprack, which could be standard or custom labware. Then the lower-level
functions have to look up the tiprack definition for that URI. There are
3 ways they try to do that:

1. Get the definition from `InvariantContext.labwareEntities`.
2. Get the definition from `PipetteEntity.tiprackLabwareDef`.
3. Get it from `getAllDefinitions()[tiprackURI]`.

The problem with (1) and (2) is that the tiprack is removed from
`InvariantContext` and `PipetteEntity` when the user unselects a tiprack
from the pipette, but a step form can still refer to the deleted
tiprack's URI. The problem with (3) is that `getAllDefinitions()` only
contains standard labware, and doesn't work for custom tipracks.

We could solve this problem if any function could look up a tiprack
definition whenever it needed to, whether the tiprack is custom or not.
But PD does not make that easy. The custom tiprack definitions are
available in `BaseState`, but most functions only have access to the
`InvariantContext`, which does **not** have all the custom labware
definitions.

I think the best approach is to **hydrate the tiprack definition into
the `HydratedMoveLiquidFormData/HydratedMixFormData`**. So this PR
changes `.tipRack` from a URI string into a tiprack labware definition.
Then any function using `Hydrated*FormData` would have the tiprack
definition without having to implement its own lookup.

(There is a precedent for this: we do the same thing when we hydrate
`.pipette` in `Hydrated*FormData`, turning the field from a string into
a `PipetteEntity`.)

In order to hydrate the tiprack definition into
`Hydrated*FormData.tipRack`, we need to pass the labware definitions
down to the hydration functions, which this PR does by calling the
`getLabwareDefsByURI()` selector.

## Test Plan and Hands on Testing

Tested PD by hand locally.
Ran all the CI tests.

## Risk assessment

Low-ish. This PR doesn't change any functionality yet, it just hydrates
the tiprack definition into `Hydrated*FormData`. An upcoming PR will
change the low-level functions to use the hydrated tiprack definition in
`Hydrated*FormData.tipRack`.