refactor(app,components): more renaming getAllLabwareDefs() -> getAllDefinitions() [cherrypick] (#20044)
# Overview

This is a cherry-pick of PR #19916. In that PR, I said that I would only
apply that change to `edge`, and leave the PD branch alone because
`chore_release-pd-8.6.1` was frozen and about to be released, and the
files with the obsolete call were not used by the PD CI tests at all.

Well, it turns out that a change I want to make to the PD branch now
does trigger the `hardware-sim/` tests for some reason:
https://github.com/Opentrons/opentrons/actions/runs/19088629372/job/54534570318

So now I need to cherrypick this change back into the PD branch.

## Test Plan and Hands on Testing

CI really really hard.

## Risk assessment

Low-ish.

(cherry picked from commit 04e051df4b06c66e32ae5b34782e670f2746b5f9)