refactor(protocol-designer): getMaxConditioningVolume() takes tiprackDef instead of URI (#20039)
# Overview

The `getMaxConditioningVolume()` function previously took a tiprack URI
and tried to look up the tiprack definition in `labwareEntities`.
However, that could be unreliable if the tiprack URI refers to a tiprack
that has been removed from the deck, because the tiprack would no longer
be in `labwareEntities`.

Now that the tiprack definition is available in the hydrated form, this
PR changes `getMaxConditioningVolume()` to just take in the tiprack
definition directly instead of trying to look it up in
`labwareEntities`.

## Test Plan and Hands on Testing

Smoke tested PD running locally.
Ran `yarn vitest --silent=false step-generation/ protocol-designer/ app/`.
Will run CI tests.

## Risk assessment

Low-ish.