refactor(protocol-designer): merge getMatchingTipLiquidSpecs() and getMatchingTipLiquidSpecsFromSpec() (#20048)
# Overview

The last one!

Back in PR #19041, @ncdiehl introduced `getMatchingTipLiquidSpecsFromSpec()`,
which was almost an exact a copy of `getMatchingTipLiquidSpecs()`
except that it looked up the tiprack definition differently, because
Nick's `getMatchingTipLiquidSpecsFromSpec()` did not have access to the
`PipetteEntity.tiprackLabwareDef`.

After the recent refactorings, the caller passes in the tiprack
definition into both `getMatchingTipLiquidSpecs()` and
`getMatchingTipLiquidSpecsFromSpec()`, so now there is no difference
between the 2 functions anymore.

So this PR merges the 2 functions together and eliminates
`getMatchingTipLiquidSpecsFromSpec()`.

## Test Plan and Hands on Testing

Run CI tests.

## Risk assessment

Low.