chore: switch to fully external buildroot tree (#20013)
This is the monorepo side of switching to having all our buildroot
changes managed in a so-called "external tree", as opposed to just
committed to a fork of the upstream buildroot repo. The changes here are
not compatible with the current kind of setup.

In this repo, all that has to be done is switching the package builders
to use our backports of future hatch tooling from upstream. When we
update buildroot, we'll be able to switch this back.

For this PR and its buildroot partner, there should be no functional
difference on a running OT-2 before and after the change - this does not
include any package upgrades or behavior changes. It's all about how we
store our customizations and how we build.

This is the monorepo side of
https://github.com/Opentrons/buildroot/pull/266 and isn't really
reviewable without that context.

## testing
- [x] put the builds of this on an ot-2 and see if it works