Docs stacker retrofit sn revision (#20058)
# Overview

Text for the Stacker upgrade kit is causing some confusion in the field.
This change strengthens language in the Flex and Stacker manuals that
indicates the upgrade kit is _required_ if a Flex has A10 or A20 in the
serial number, with or without the HEPA/UV module.

Robots with **A30** in their serial numbers do not require this upgrade.

**Sandbox:**
- Stacker manual:
https://sandbox.docs.opentrons.com/docs-stacker-retrofit-sn-revision/stacker/compliance/#flex-stacker-and-hepauv-compatibility

- Flex manual:
https://sandbox.docs.opentrons.com/docs-stacker-retrofit-sn-revision/flex/modules/stacker/#module-compatibility

RTC-893

## Test Plan and Hands on Testing

Read the wordz for clarity and accuracy.

## Changelog

Revises text here:

- `flex/docs/modules/stacker.md#module-compatibility`
- `stacker/docs/compliance#flex-stacker-and-hepa/uv-compatibility`

This change will also update the [retrofit kit
guide](https://docs.google.com/document/d/1UxfxRWObhTvgp34Y_7ZcCkdmS3bi65ohzewiHTC0QjY/edit?usp=sharing)
with the more comprehensive text used by the Stacker manual after this
PR is reviewed and approved.

## Review requests

Does this change make it clear that the retrofit kit is required for
older model robots, even without the HEPA module?

## Risk assessment

Medium: can cause field service confusion.

---------

Co-authored-by: Ed Cormany <edward.cormany@opentrons.com>