fix(app, app-shell): Refresh secondary window stream on new run (#20087)
Currently, we inject the `runId` into the secondary window, and that
`runId` serves as the basis for determining whether or not we should
render the stream. However, if a user keeps the secondary window open
when creating a new run on the same robot, the stream still reflects the
old run.

To fix, we check if this robot has a current run, and use that as the
basis for rendering the livestream. All of this works as expected
largely because we block the "show livestream" button for historical
runs.