fix(app): Conditionally change check preferences camera text color  (#20082)
# Overview

Changes the `check_preferences` text to yellow if the image storage is
almost full, red if the camera is disabled but needed for the run, grey
if the camera settings are is ready to be confirmed, and green if the
camera settings have been confirmed.

## Test Plan and Hands on Testing

- smoke tested on robot


https://github.com/user-attachments/assets/144dce45-58f7-4da8-b9d0-c7a819ca2176

## Changelog

- an additional param was added to check the storage info
- camera set up is now being viewed as a module set up type to allow for
checking of camera enablement and storage

## Review requests

- What do you need from reviewers to feel confident this PR is ready to
merge?
- Ask questions.

## Risk assessment
- lowish
- may have accidentally changed other text formatting

Closes
RQA-4828