feat(protocol-designer): handle full tip selection (#20101)
This PR adds the correct state for the SelectTips component for when all
pickup selections have been made. Specifically, it wires up the success
Chip component in the top right of the component above the deck map, and
adds an error state for all non-deselect hovers if there are no pickups
remaining.