fix(components): don't forward our `showBorder` prop to <Box> and <div> (#20114)
# Overview

The console is cluttered up with warnings of:
```
Warning: React does not recognize the `showBorder` prop on a DOM element.
If you intentionally want it to appear in the DOM as a custom attribute,
spell it as lowercase `showborder` instead. If you accidentally passed it
from a parent component, remove it from the DOM element.
```

`showBorder` (and `isZoomed`) is a custom property we invented for
`<StyledBox>`. We need to prevent it from being forwarded down to the
`<div>`.

## Test Plan and Hands on Testing

Tried it in PD running locally.

## Review requests

I haven't touched styled components before, so I don't really know what
I'm doing.

Also, it seems like a lot of the existing code says:
```
{ shouldForwardProp: isntStyleProp }
```
Isn't that the opposite of what we want? This is saying that if
**isn't** one of the native style properties, then we **do** forward it.
Don't we want this to be `{ shouldForwardProp: isStyleProp }` instead?

## Risk assessment

Low.