fix(build): build py dists before deploy (#20108)
The old opentrons deploy actions would build the distributables as well
as deploy them; now that we're using the pypi-maintained deploy actions,
we need to build them explicitly (we can't use the ones we already built
because that happened in a different job aka vm).

## Testing
Unfortunately there's not much we can do here, but this is already
broken and this is pretty straightforward.