fix(docs): swap comment on set_offset() and calibrated_offset() (#20123)
# Overview

Previously, the `Labware.set_offset()` function claimed that
```
The offset is an x, y, z vector in deck coordinates (see :ref:`protocol-api-deck-coords`).
```
This is misleading because a deck coordinate of `(0, 0, 0)` means the
front left corner of the deck, but `set_offset(0, 0, 0)` would **not**
put the labware on the front left corner of the deck.

Then there's also the `Labware.calibrated_offset` property, which
**does** return deck coordinates, but is not documented as doing so.

This PR swaps the deck coordinates comment from `Labware.set_offset()`
to `Labware.calibrated_offset`.

## Test Plan and Hands on Testing

No change to functionality at all, just a docstring change.

## Risk assessment

Low.