fix(hardware): fix divide by zero error on some plunger moves. (#20146)
# Overview
Apparently there are some rare instances in drop tip that can result in
a malformed target where the plunger axes gets included in the target
but ends up with a 0 unit vector.

I believe this has to be a result of numpy calculating a very very tiny
float, for this move and when it gets used as a python native float it
turns into 0.0

Added a check and unit test.