fix(app-shell, app-shell-odd): Report window type on hard refresh (#20156)
Closes RQA-4773

We report the window type when the app initially renders in App.tsx, and this window type determines the specific render tree for a given renderer process. However, we weren't reporting this when a user hard refreshes an app, because the report only happens on initial render. To fix, we can report the event on did-finish-load continuously after the initial window load.