fix(api): preserve critical points in tracking end
When we calculate the move target for the tracking gantry/z move, we
need to take into account any critical point overrides from things like
multichannels needing to be centered on 1-channel reservoirs.

Luckily, this is low-risk since the functions the tracking mechanism in
hardware control uses to calculate the position take a critical point,
and the engine knows how to calculate critical points, so we're not
actually making any decisions here - just plumbing additional data
through.

Closes RQA-4875