fix(api): preserve critical points in tracking end (#20163)
When we calculate the move target for the tracking gantry/z move, we
need to take into account any critical point overrides from things like
multichannels needing to be centered on 1-channel reservoirs.

Luckily, this is low-risk since the functions the tracking mechanism in
hardware control uses to calculate the position take a critical point,
and the engine knows how to calculate critical points, so we're not
actually making any decisions here - just plumbing additional data
through.

## Testing
- Do a tracking aspirate and dispense using any multiple pipette layout
(8-channel pipette in all-tips mode, 96-channel pipette in any mode but
single) on a 1-well reservoir, or any column layout (8-channel pipette
in all-tips mode or 96-channel pipette in column mode on one column or
all-tips mode on well A1) on any kind of reservoir. The pipette should
not move in X or Y while aspirating or dispensing.

Closes RQA-4875